+++ src/Item/index.types.ts
@@ -1,4 +1,4 @@
 // Types
-import type { GridProps } from '@mui/material'
+import type { Grid2Props as GridProps } from '@mui/material'
 
 export type SpinoramaItemProps = GridProps & {}