+++ src/Thumbnail/index.types.ts
@@ -1,5 +1,5 @@
 // Types
-import type { GridProps } from '@mui/material'
+import type { Grid2Props as GridProps } from '@mui/material'
 
 export type SpinoramaThumbnailProps = GridProps & {
 	/**